AssignStation + MatchView + enabled Background
Almost done, timers/local push notifications need to be programmed and
the sorting needs to be programmed

Added current_station to Match data

MatchView Assign Station Button title should change correctly now
depending if there is a corresponding station to the match. The timer
text needs to be input as well after the timers are made in the
AssignStationView.

Background has been enabled. This is so timers can function in the
background and a push notification can be sent even while the user is
not in the application or AssignStation view. This functionality still
needs to be programmed.