notification and notification-alertview update 1
this is the first update. Notifications outside of app are working as
intended, just need to find a way to modify the view it opens up to if
possible or just set it to the home view when pressed

notification alerts for when in app have been solved. Just need to
program.

yay